Factoring, Renaming files in English, reorganizing files, updating README
