+++ README.md
@@ -1,1 +1,89 @@
-# aventuregame
+<a name="readme-top"></a>
+
+![GitHub contributors](https://img.shields.io/github/contributors/xTsuKiZox/aventuregame?color=0d0&style=for-the-badge)
+![GitHub release](https://img.shields.io/github/v/release/xTsuKiZox/aventuregame?style=for-the-badge)
+![GitHub watchers](https://img.shields.io/github/watchers/xTsuKiZox/aventuregame?style=for-the-badge)
+![GitHub Repo stars](https://img.shields.io/github/stars/xTsuKiZox/aventuregame?color=%23fa0&style=for-the-badge)
+![GitHub issues](https://img.shields.io/github/issues/xTsuKiZox/aventuregame?style=for-the-badge)
+
+
+
+<!-- LOGO -->
+<br />
+<div align="center">
+    
+  <!-- <img src="https://dejpknyizje2n.cloudfront.net/svgcustom/clipart/preview/vibrant-sunflower-sticker-30284-550x550.png" alt="Logo" width="80" height="80"> -->
+
+  <h3 align="center">TEXTUAL VIDEO GAME</h3>
+
+  <p align="center">
+    Welcome to the project : Chrono PARADOX !
+    <br>
+    You are currently reading the documentation, happy reading :)
+  </p>
+</div>
+
+
+
+<!-- TABLE  -->
+<details>
+  <summary>Table of Contents :</summary>
+  <ol>
+    <li>
+      <a href="#about-the-project">About The Project</a>
+      <ul>
+        <li><a href="#built-with">Built With</a></li>
+      </ul>
+    </li>
+    <li>
+      <a href="#getting-started">Getting Started</a>
+      <ul>
+        <li><a href="#prerequisites">Prerequisites</a></li>
+        <li><a href="#installation">Installation</a></li>
+      </ul>
+    </li>
+  </ol>
+</details>
+
+
+
+<!-- ABOUT THE PROJECT -->
+## About The Project
+
+
+Chrono PARADOX is a text adventure game created from scratch by 3 developers !
+
+This is a text adventure game: Chrono Paradox produced by our studio Sunflower
+
+*Note : This project is a student project*
+
+<p align="right">(<a href="#readme-top">Back to top !</a>)</p>
+
+
+
+### Built With
+
+This section lists the main frame used to start your project. We were free to choose on which language we will program, we chose to program our game on Python 
+
+* ![Python](https://img.shields.io/badge/python-3670A0?style=for-the-badge&logo=python&logoColor=ffdd54)
+
+<p align="right">(<a href="#readme-top">Back to top !</a>)</p>
+
+
+
+<!-- GETTING STARTED -->
+## Getting Started
+
+To be able to play Chrono Paradox, you must stick to it as follows :
+
+
+### Installation
+
+_Below is how you can how to install Chrono Paradox . You have to_ :
+
+1. Download the repository,
+2. From your computer launched it via a terminal
+3. Enjoy the game !
+
+<p align="right">(<a href="#readme-top">Back to top !</a>)</p>
+<!--  -->
