--- caracteristiques_personnage.xml
@@ -1,6 +1,0 @@
-<personnage>
-    <nom>VotreNom</nom>
-    <niveau>1</niveau>
-    <vie>100</vie>
-    <force>10</force>
-</personnage>
