+++ game/character_characteristics.xml
@@ -1,0 +1,6 @@
+<personnage>
+    <nom>VotreNom</nom>
+    <niveau>1</niveau>
+    <vie>100</vie>
+    <force>10</force>
+</personnage>
